pyproject.toml
[tool.poetry]
name = "zero-deep1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.0.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
